+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `databaseSettings` parameter for `postSaveHook()` method
+
+### Changed
+- `Client` Model now set the new clients databaseKey from config
+
+### Fixed
+- `Client Create API` `postSaveHook()` client codes parameter
 
 ## [1.1.0] - 2020-05-19
 ### Removed